Improve the loader and the floating point module
A new program, written in Forth for Gforth, patches the source of the
BASIC loader with the current load and entry addresses compiled in the
kernel. So far those addresses were hardcoded in the source of the
loader. Now the loader is automatically updated by Makefile when needed.

This improvement was pending, but now it's especially useful in order to
debug some words of the floating point module, changing the address of
the kernel to free more memory for the ROM calculator stack.

This tool and others under development have been included in the tools
directory.

FossilOrigin-Name: a3b9f02cc02cc0fb329651a6c5de170a20df706d88a4e7d770d8a4203c73f8b3