Fix `fmax`, `fmin` and calculator's `if`
The calculator's `if` compiled the "jump-true" calculator command, which
does a branch when calculator TOS is not zero, the opposite of what it
should be.

The bug of `if` affected `fmax` and `fmin`, but these words had to be
modified also because they used the calculator's operators `>` and `<`
directly, what causes problems (the details are in the source of the
calculator module).

FossilOrigin-Name: 1667aead41a4784edebba612133f3bc8472692ce17f16dc89bb6e4698942256d