Simplify the circular string buffer
So far the internal offset of the buffer was stored before the buffer
data. It has been moved to `>csb`, which formerly was a fake variable, a
constant that returned the address of the offset, and now is an actual
variable. This saves a cell and is more versatile. In total, ten bytes
are saved with some related changes.

FossilOrigin-Name: 87d3addfae77c441b6b54d77ad7614806d139bd23d9aef68e0f35e14c6f15c7c