Modify the structure of vocabularies
Now vocabularies don't keep the data of a word list in their body.
Instead, they create a wordlist and store only its address. This way,
`vocabulary` can be moved to the library, even if there are three
vocabularies in the kernel (`root`, `forth` and `assembler`). But the
main advantage of this change is conversion can work in both directions:
a word list identifier can be get from a vocabulary name, no matter if
the vocabulary was created with `vocabulary` or out of an existent word
list; and a vocabualary can be created from a word list, and its
structure will be identical to those created by `vocabulary`.

FossilOrigin-Name: 43c99cd8f297a039011ec6eaae9b13d757807882ec0ba6ac24c4a1cce78adda8