Split the meta tools disk into two
So far the meta tools disk contained the main library, the tests and the
benchmarks. It fitted for a 800-KiB G+DOS disk image, but not for a
636-KiB TR-DOS disk image.

Now the tests and the benchmarks are in different disks.

FossilOrigin-Name: 65b0d3a5c45889cc3557adadd73717f0254bd0a72d884adc8684f241ed5dcaf7