Fix prerequisites to build the BASIC loaders
Add the kernel to the prerequisites of the patched loaders. This fixes a
subtle bug: The patched loader was not updated the first time after a
change in the kernel, even if the change implied a symbol. This made the
system crash on TR-DOS when the `origin` address was modifed, unless
`make` was executed twice to force the loader be updated.

FossilOrigin-Name: 6eb6eb3de47da4315f35347b595586f26ea91bc8d37a2fb33096fa9e93ec816f