Halve the `stringer` and rename its main words
The `stringer` (circular string buffer9 was going to be moved at the top
of the main memory, reserving space above `limit`, so the application
can move it to data space and increase its size if needed, without
wasting its original space, which now is isolated in the kernel and can
not be returned to data space.

But any location above address $C000, the top 16 KiB of RAM, which are
shared with paged memory,  is unusable during compilation, because
string manipulation is required to build the word headers in far memory.

As a compromise solution, the size of the `stringer` has been halved to
256 bytes.

Words renamed:

- `>stringer` to `+stringer`.
- `save-string` to `>stringer`.
- `save-farstring` to `far>stringer`
- `allocate-string` to `allocate-stringer`
- `save-counted-string` to `counted>stringer`.

The new names are more consistent as a set, and also more coherent with
the general naming convention.

Besides, all mentions to "circular string buffer" have been updated to
"`stringer`".

FossilOrigin-Name: 9fb19ba089e8763fd8247a5228959931b928c7856295c8f00128d1097aa73be1