Remove `os-used`, rewrite `os-unused`, add `?os-unused`
The problem was the ROM routine called by `os-used` included the
position of the Z80 stack, which is different in Forth and in BASIC.
Therefore the calculation of `os-unused` was wrong. A simpler
calculation based on `os-stkend` and `os-ramtop` seems to work fine.

FossilOrigin-Name: c59614becdf05be75bec96d68d9d8ebe428863dba7275c032c0ebe9e7a0a606c